Move path & spanned into Control.Effect.Parser.
